% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{accelRate}
\alias{accelRate}
\title{This function estimates the external training load from accelerometry}
\usage{
accelRate(ax3)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}
}
\value{
tl Training Load
}
\description{
\code{accelRate} estimates the external training load from accelerometry
}
\seealso{
\code{\link{playerLoad}}
}
