% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{batchTraingLoad}
\alias{batchTraingLoad}
\title{This function estimates the external training load from accelerometry for all files in the given folder}
\usage{
batchTraingLoad(folder, startTime = 0, endTime = 200)
}
\arguments{
\item{folder}{Folder to the directory containing the Axivity files}

\item{startTime}{time in seconds for when to start recording}

\item{endTime}{time in seconds for when to end the recording}
}
\value{
df Training Load for all files in the folder
}
\description{
\code{batchTraingLoad} estimates the external training load from accelerometry for all files in the given folder
}
\seealso{
\code{\link{playerLoad}}
}
