% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{batchTRIM}
\alias{batchTRIM}
\title{This function estimates TRIM from garmin tcx for all files in the given folder}
\usage{
batchTRIM(folder, subd = 0)
}
\arguments{
\item{folder}{Folder to the directory containing the Axivity files}

\item{subd}{data.fram containing the subject ID, restHR, maxHR and gender information. ID must be the same as the file name}
}
\value{
df TRIM for all files in the folder
}
\description{
\code{batchTRIM} estimates TRIM from accelerometry for all files in the given folder
}
\seealso{
\code{\link{bTRIM,eTRIM,eTRIMSpecial}}
}
