% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{enmo}
\alias{enmo}
\title{This function estimates Euclidean Norm minus one from accelerometry}
\usage{
enmo(ax3, epoch = 5)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}

\item{epoch}{Epoch length in seconds - Default is 5 seconds}
}
\value{
enmo data
}
\description{
\code{enmo} estimates ENMO from accelerometry
}
\seealso{
\code{\link{playerLoad}}
}
