% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{eTRIM}
\alias{eTRIM}
\title{This function estimates the intensity distribution Training Impulse}
\usage{
eTRIM(tcxTable, restHR, maxHR, gender)
}
\arguments{
\item{tcxTable}{Heart Rate data from tcx file}

\item{restHR}{Resting heart rate}

\item{maxHR}{Maximum Heart rate}

\item{gender}{Gender of the subject}
}
\value{
tl Training Load
}
\description{
\code{bTRIM} Estimates the intensity distribution Training Impulse
}
\seealso{
\code{\link{bTRIM}}
}
