% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{eTRIMspecial}
\alias{eTRIMspecial}
\title{This function estimates the Edwards intensity distribution Training Impulse}
\usage{
eTRIMspecial(tcxTable, restHR, maxHR, gender)
}
\arguments{
\item{tcxTable}{Heart Rate data from tcx file}

\item{restHR}{Resting heart rate}

\item{maxHR}{Maximum Heart rate}

\item{gender}{Gender of the subject 0: Male 1: Female}
}
\value{
tl Training Load
}
\description{
\code{eTRIMspecial} Estimates the Edwards intensity distribution Training Impulse
}
\seealso{
\code{\link{bTRIM}}
}
