% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{mad}
\alias{mad}
\title{This function estimates Mean Absolut deviation from accelerometry}
\usage{
mad(ax3, epoch = 5)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}

\item{epoch}{Epoch length in seconds - Default is 5 seconds}
}
\value{
mad data
}
\description{
\code{mad} estimates MAD from accelerometry
}
\seealso{
\code{\link{playerLoad}}
}
