% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{playerLoad2}
\alias{playerLoad2}
\title{This function estimates the external training load from accelerometry}
\usage{
playerLoad2(ax3)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}
}
\value{
tl Training Load
}
\description{
\code{playerLoad2} estimates the external training load from accelerometry
}
\seealso{
\code{\link{accelRate}}
}
