% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{playerLoadSUM}
\alias{playerLoadSUM}
\title{This function estimates the external training load from accelerometry}
\usage{
playerLoadSUM(ax3)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}
}
\value{
tl Training Load
}
\description{
\code{playerLoadSUM} estimates the external training load from accelerometry
}
\seealso{
\code{\link{accelRate}}
}
