% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{velocityLoad}
\alias{velocityLoad}
\title{This function estimates the external training load from accelerometry}
\usage{
velocityLoad(ax3)
}
\arguments{
\item{ax3}{AX3 Accelerometry data}
}
\value{
tl Training Load
}
\description{
\code{velocityLoad} estimates the external training load from accelerometry using ENMO and filtering the data
}
\seealso{
\code{\link{playerLoad}}
}
