% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingLoad.R
\name{bTRIM}
\alias{bTRIM}
\title{This function estimates the Banister Training Impulse (TRIM)}
\usage{
bTRIM(tcxTable, restHR, maxHR, gender)
}
\arguments{
\item{tcxTable}{Heart Rate data from tcx file}

\item{restHR}{Resting heart rate}

\item{maxHR}{Maximum Heart rate}

\item{gender}{Gender of the subject 0: Male 1: Female}
}
\value{
tl Training Load
}
\description{
\code{bTRIM} Estimates the Banister Training Impulse (TRIM)
}
\seealso{
\code{\link{eTRIM}}
}
